fixed PR coments
